UP-1653 Minor organization for some new types
