+++ src/main/scala/com/ubirch/models/PublicKey.scala
@@ -40,7 +40,7 @@ object PublicKeyInfo {
   final val VALID_NOT_AFTER = "validNotAfter"
   final val VALID_NOT_BEFORE = "validNotBefore"
 
-  def fixValuesFomMsgPack(jv: JValue) = {
+  def fixValuesFomMsgPack(jv: JValue): JValue = {
     def formatter(time: Long) = DateUtil.ISOFormatter.print(time.toLong * 1000)
     jv.mapField {
       case (x @ PublicKeyInfo.ALGORITHM, JString(value)) => (x, JString(new String(Base64.getDecoder.decode(value))))
@@ -52,7 +52,7 @@ object PublicKeyInfo {
     }
   }
 
-  def checkPubKeyId(jv: JValue) = {
+  def checkPubKeyId(jv: JValue): JValue = {
     val pk = jv.findField {
       case (PublicKeyInfo.PUB_KEY, _) => true
       case _ => false
